update cld tests
