+++ tests/experiments/cld/test_cld_kinematics.py
